refactor: create <OurTeamSection />
